requirements.txt
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.1.0,<2.0.0
tensorflow>=2.10.0,<3.0.0  # Use tensorflow-cpu if you don't have GPU support
matplotlib>=3.5.0,<4.0.0
seaborn>=0.12.0,<1.0.0
PyYAML>=6.0,<7.0.0
ipykernel>=6.0.0,<7.0.0 # For notebooks
requests>=2.28.0,<3.0.0 # If you add a download script